Add Defaults to support additional lexical scopes
This patch adds a Defaults and its support in Pointer.Get. When Defaults
is received Pointer.Get will look for the target in the first element of
Defaults and return it if it is found or go to the next element.

This makes it possible to support lexical scopes for the `any` and `all`
loop constructs in go-bexpr and still have meaningful error messages when
something is not found.